Fix flaky Deno unit tests caused by Vitest fork and worker thread issues.
Isolate getUserSharedBuffer callback tests in their own file, run Bun/Deno
tests sequentially, and centralize worker teardown so fork workers exit cleanly.
