+++ test/lock.test.ts
@@ -1,4 +1,4 @@
-import { createWorkerBootstrapScript, dbRunner } from './lib/util.js';
+import { createWorkerBootstrapScript, dbRunner, terminateWorker } from './lib/util.js';
 import { setTimeout as delay } from 'node:timers/promises';
 import { Worker } from 'node:worker_threads';
 import { describe, expect, it, vi } from 'vitest';
@@ -143,6 +143,7 @@ describe('Lock', () => {
 					expect(db.tryLock('foo', () => resolve())).toBe(false);
 					worker.postMessage({ close: true });
 				});
+				await terminateWorker(worker);
 			}));
 	});
 
@@ -295,22 +296,7 @@ describe('Lock', () => {
 				expect(spy).toHaveBeenCalledTimes(3);
 
 				worker.postMessage({ close: true });
-
-				if (process.versions.deno) {
-					// there is something buggy with Deno where calling `await delay(100)` freezes the
-					// process, but advancing a microtask seems to unfreeze it
-					await new Promise<void>((resolve) => {
-						const timer = setTimeout(() => {
-							worker.terminate();
-							resolve();
-						}, 100);
-
-						worker.on('exit', () => {
-							clearTimeout(timer);
-							resolve();
-						});
-					});
-				}
+				await terminateWorker(worker);
 			}));
 	});
 });
